+++ rust-doomgeneric/src/p_map.rs
@@ -1086,7 +1086,7 @@ pub unsafe fn P_ChangeSector(
     return nofit != 0;
 }
 unsafe fn SpechitOverrun(mut ld: *mut line_t) {
-    static mut baseaddr: u32 = 0 as u32;
+    static mut baseaddr: u32 = 0;
     let mut addr: u32 = 0;
     if baseaddr == 0 as u32 {
         let mut p: i32 = 0;
